add filter search

+++ app/dept_sys/routes.py
@@ -32,66 +32,77 @@ def after_request(response):									#security
 
 @dept_sysbp.route('/check_dept_json',methods=['GET','POST'])
 def checkdeptjson():
-	deptrec = [dept.to_json() for dept in Department.query.all()]
-	response = { 'records': deptrec }
-	return jsonify(response)
+	try:
+		deptrec = [dept.to_json() for dept in Department.query.all()]
+		response = { 'records': deptrec }
+		return jsonify(response)
+	except:
+		return {'status':'data fetch failed'}
 
 @dept_sysbp.route('/add_dept_json',methods=['POST'])
 def adddeptjson():
 	jsdat = request.json
 	check_dept = Department.query.filter_by(ID = jsdat.get('id')).all()
 
-	if  check_dept:
-		return jsonify({ 'error' : 'dept already in database'})
+	if check_dept:
+		return jsonify({ 'status' : 'dept already in database'})
 
 	if jsdat.get('id') == '' or jsdat.get('id') is None:
-		return jsonify({ 'error' : 'bad info'})
+		return jsonify({ 'status' : 'empty dept id'})
 	if jsdat.get('name') == '' or jsdat.get('name') is None:
-		return jsonify({ 'error' : 'bad info'})
+		return jsonify({ 'status' : 'empty dept name'})
 
 	dept = Department.from_json(jsdat)
 	if dept is None:
-		return jsonify({ 'error' : 'bad info'})
+		return jsonify({ 'status' : 'can\'t create department record'})
 
-	db.session.add(dept)
-	db.session.commit()
-	return jsonify({ 'status' : 'success'})
+	try:
+		db.session.add(dept)
+		db.session.commit()
+		return jsonify({ 'status' : 'Dept add success'})
+	except:
+		return jsonify({ 'status' : 'Dept add fail'})
 
 @dept_sysbp.route('/modify_dept_json',methods=['POST'])
 def modifydeptjson():
 	oldjs = request.json['old']
 	newjs = request.json['new']
 	
-	dept = Department.query.filter_by(ID=oldjs.get('id')).first_or_404()
+	dept = Department.query.filter_by(ID=oldjs.get('id'),name=oldjs.get('name')).first_or_404()
 	if dept is None:
-		return jsonify({ 'error' : 'bad info'})
-
-	dept = Department.query.filter_by(name=oldjs.get('name')).first_or_404()
-	if dept is None:
-		return jsonify({ 'error' : 'bad info'})
+		return jsonify({ 'status' : 'No such Department as {}'.format(newjs.get('name'))})
 
 	if newjs.get('id') == '' or newjs.get('id') is None:
-		return jsonify({ 'error' : 'bad info'})
+		return jsonify({ 'status' : 'empty dept id'})
+
 	if newjs.get('name') == '' or newjs.get('name') is None:
-		return jsonify({ 'error' : 'bad info'})
+		return jsonify({ 'status' : 'empty dept name'})
 
 	check_dept = Department.query.filter_by(ID = newjs.get('id')).all()
 
 	if not check_dept:
-		return jsonify({ 'error' : 'dept already in database'})
+		return jsonify({ 'status' : 'Dept {} already in database'.format(newjs.get('id'))})
 
 	dept.ID = newjs.get('id') or dept.ID
 	dept.name = newjs.get('name') or dept.name
 
-	db.session.commit()
-	return jsonify({ 'status' : 'success'})
+	try:
+		db.session.commit()
+		return jsonify({ 'status' : 'Dept modify success'})
+	except:
+		return jsonify({ 'status' : 'Dept modify fail'})
 
 @dept_sysbp.route('/delete_dept_json',methods=['POST'])
 def deldeptjson():
+	if not request.get_data('id'):
+		return jsonify({ 'status' : 'No id given'})
+
 	dept = Department.query.filter_by(ID=request.get_data('id')).all()
 	if not check_dept:
-		return jsonify({ 'error' : 'dept does not exist'})
-
-	db.session.delete(dept)
-	db.session.commit()
-	return jsonify({ 'status' : 'success'})
+		return jsonify({ 'status' : 'No such department in database'})
+	try:
+		db.session.delete(dept)
+		db.session.commit()
+		return jsonify({ 'status' : 'Dept deletion success'})
+	except:
+		return jsonify({ 'status' : 'Dept deletion failed'})
